Param类删除getParams()方法
