+++ app/src/main/java/com/example/httpsender/param/PostEncryptJsonParam.kt
@@ -2,10 +2,10 @@ package com.example.httpsender.param
 
 
 import okhttp3.RequestBody
-import org.json.JSONObject
 import rxhttp.wrapper.annotation.Param
 import rxhttp.wrapper.param.JsonParam
 import rxhttp.wrapper.param.Method
+import rxhttp.wrapper.utils.GsonUtil
 
 /**
  * User: ljx
@@ -21,7 +21,7 @@ class PostEncryptJsonParam(url: String) : JsonParam(url, Method.POST) {
     override fun getRequestBody(): RequestBody {
         //我们要发送Post请求，参数以加密后的json形式发出
         //第一步，将参数转换为Json字符串
-        val json = JSONObject(params).toString()
+        val json = if (params == null) "" else GsonUtil.toJson(params)
         //第二步，加密
         val encryptByte = encrypt(json, "RxHttp")
         //第三部，创建RequestBody并返回
