+++ rxhttp/src/main/java/rxhttp/wrapper/param/NoBodyParam.java
@@ -1,5 +1,9 @@
 package rxhttp.wrapper.param;
 
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import io.reactivex.annotations.NonNull;
 import okhttp3.HttpUrl;
 import okhttp3.RequestBody;
 import rxhttp.wrapper.utils.BuildUtil;
@@ -12,6 +16,7 @@ import rxhttp.wrapper.utils.BuildUtil;
  */
 public class NoBodyParam extends AbstractParam<NoBodyParam> {
 
+    private Map<String, Object> mParam; //请求参数
     /**
      * @param url    请求路径
      * @param method {@link Method#GET,Method#HEAD,Method#DELETE}
@@ -27,11 +32,32 @@ public class NoBodyParam extends AbstractParam<NoBodyParam> {
 
     @Override
     public HttpUrl getHttpUrl() {
-        return BuildUtil.getHttpUrl(getSimpleUrl(), getParams());
+        return BuildUtil.getHttpUrl(getSimpleUrl(), mParam);
+    }
+
+    @Override
+    public NoBodyParam add(String key, Object value) {
+        if (value == null) value = "";
+        Map<String, Object> param = mParam;
+        if (param == null) {
+            param = mParam = new LinkedHashMap<>();
+        }
+        param.put(key, value);
+        return this;
     }
 
     @Override
     public final RequestBody getRequestBody() {
         return null;
+    }
+
+    @NonNull
+    public Map<String, Object> getParams() {
+        return mParam;
+    }
+
+    @Override
+    public String toString() {
+        return getUrl();
     }
 }
